setColorScheme()

Sets color scheme for a chart and the category names (optional)

Usage

setColorScheme(myIO, colorScheme = NULL, setCategories = NULL)

Arguments

myIO

an htmlwidget object created by the myIO() function

colorScheme

a vector of colors in the order you want them used

setCategories

an optional vector of names that will be mapped to the corresponding color in the colorScheme

Value

the same myIO object

Examples

# Set a custom color scheme
myIO() |> setColorScheme(colorScheme = list("red", "blue", "green"))

# Set colors with category labels
myIO() |> setColorScheme(
  colorScheme = list("steelblue", "orange"),
  setCategories = c("Group A", "Group B")
)