setReferenceLines()

Sets x and y reference lines

Usage

setReferenceLines(myIO, xRef = 0, yRef = 0)

Arguments

myIO

an htmlwidget object created by the myIO() function

xRef

a list of the reference line value of x

yRef

a list of the reference line value of y

Value

the same myIO object

Examples

# Add reference lines at x=5 and y=20
myIO() |> setReferenceLines(xRef = 5, yRef = 20)