myIO

myIO R + d3.js project

Usage

myIO(data = NULL, width = "100%", height = "400px", elementId = NULL)

Arguments

data

an optional point of entry for the data frame or vector

width

a string of either pixel width or a percentage width

height

a string of pixel height

elementId

a unique id for the htmlwidget object

Value

the same myIO object